[step2] refactor package
