+++ src/main/java/model/http/HttpVersion.java
@@ -1,4 +1,4 @@
-package webserver;
+package model.http;
 
 import java.util.Arrays;
 import java.util.Optional;
