+++ src/main/java/model/http/Query.java
@@ -1,4 +1,4 @@
-package webserver;
+package model.http;
 
 import java.util.Arrays;
 import java.util.List;
