+++ src/test/java/model/http/PathTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package model.http;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
